fixed bugs with average

+++ Sprint5/FinalProgrammingAssignment-SearchEngine/hashtableadapt.h
@@ -10,7 +10,7 @@ public:
     ~HashtableAdapt();
     int& getSize();
     int& getNumDocsParsed();
-    int& getAverageWordsPerFile();
+    double& getAverageWordsPerFile();
     Word* insert(Word&);
     void makeEmpty();
     bool isEmpty() const;
@@ -21,7 +21,7 @@ public:
 private:
     Hashtable<string,Word> theIndex;
     int numDocsParsed;
-    int averageWordsPerFile;
+    double averageWordsPerFile;
 };
 
 #endif // HASHTABLEADAPT_H
